+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### New (Unreleased)
  - Added custom exception classes
+ - Added disable_proxy client option
 
 ### v1.1.0
  - Added ManagerAccountHelper (to allow setting user permissions)