+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ### New (Unreleased)
+
+### v1.2.0
  - Added custom exception classes
  - Added disable_proxy client option
  - Added CLI
  - Support environment variables
+ - New BIOS and ComputerSystem helper methods (deprecated old ones)
 
 ### v1.1.0
  - Added ManagerAccountHelper (to allow setting user permissions)